pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "continual-llm"
version = "0.1.0"
description = "A real-time adaptive LLM with continual learning capabilities"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["machine learning", "deep learning", "continual learning", "language models", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "regex>=2023.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "isort>=5.12.0",
]
data = [
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "pandas>=2.0.0",
]
training = [
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
]
all = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "pandas>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/Custom_ML_Agent"
Documentation = "https://github.com/yourusername/Custom_ML_Agent#readme"
Repository = "https://github.com/yourusername/Custom_ML_Agent"
"Bug Tracker" = "https://github.com/yourusername/Custom_ML_Agent/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["**/*.yaml", "**/*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
